BoardConfig: Add a default TARGET_BOOTLOADER_BOARD_NAME
There is no need to break the build when a TARGET_BOOTLOADER_BOARD_NAME can not be derived from TARGET_PRODUCT.
Simply set a valid default value, and warn the user about the issue.
This is useful to guarantee compatibility with projects based on Sony's Open Devices that change TARGET_PRODUCT to fit their needs.

Change-Id: I0366d983232e6fc505068dd0f7fce9dbf9ba7e94